+++ actionfx-core/src/main/java/com/github/actionfx/core/view/AbstractView.java
@@ -283,6 +283,7 @@ public abstract class AbstractView implements View {
 		return stylesheets;
 	}
 
+	@Override
 	public ResourceBundle getResourceBundle() {
 		return resourceBundle;
 	}