+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ## 1.3.0 (13.06.2021)
 - Added attribute 'filterPredicateProperty' to annotation @AFXUseFilteredList in order to bind a predicate to the filtered list directly
+- Extended ControllerWrapper class to retrieve the java.util.ResourceBundle from the controller instance
 
 ## 1.2.0 (12.06.2021)
 - Annotation @AFXOnAction now supports asynchronous method invocation via attribute 'async'.